Convert Preference to an interface to allow easy mocking for unit tests
